# Speed up RSpec boot time by 30%