--- a/Gemfile
+++ b/Gemfile
@@ -216,7 +216,7 @@ group :test do
   # and other niceties
   gem 'test-prof', '~> 1.0.0'
 
-  gem 'database_cleaner', '~> 2.0'
+  gem 'database_cleaner', '~> 2.0' # only useful for legacy_spec
   gem 'rack_session_access'
   gem 'rspec', '~> 3.11.0'
   # also add to development group, so "spec" rake task gets loaded