By passing `pre_count: true` to `DatabaseCleaner`, tables will be checked for existing rows before truncating. As we are using transactional strategy, tables are rolled back after the test suite and are empty, so pre counting the rows to prevent truncation saves time.

As tables are truncated only once at the beginning of the test suite, this saves boot time. When running
`time rspec spec/services/work_packages/shared/working_days_spec.rb:47`:

  * `pre_count: false` => 10,370 seconds total
    > Finished in 3.1 seconds (files took 6.14 seconds to load)

  * `pre_count: true`  =>  7,365 seconds total
    > Finished in 0.06816 seconds (files took 6.17 seconds to load)