@oliverguenther 2022-10-05
I'm wondering, do we really need DatabaseCleaner with transactions being in place? I can't think of a reason why unless for legacy specs