@cbliard 2022-10-05
I was wondering the same. I guess it is still useful if RSpec test get killed for some reason and the database is left with some data in it. But in that case I would assume the transaction is rolled back as well.

Maybe we could try removing it.

I'll look into why the tests are failing.