@cbliard 2022-10-06
> But speaking of getting rid of an unnecessary step, can you please add a comment to the Gemfile triggering us to remove `database_cleaner` once the legacy specs are removed?

Yep done.